fix: rebase and fixed tests
